--- mocks/get.yaml
@@ -1,12 +1,0 @@
-when:
-  method: GET
-  path: /
-  header:
-    - name: Test
-      value: test
-then:
-  status: 200
-  header:
-    - name: content-type
-      value: text/html
-  body: "Test Successful"
