--- mocks/get_api_key_header.yaml
@@ -1,12 +1,0 @@
-when:
-  method: GET
-  path: /test_api_key
-  header:
-    - name: X-TYPESENSE-API-KEY
-      value: VerySecretKey
-then:
-  status: 200
-  header:
-    - name: content-type
-      value: text/html
-  body: "Test with api key successful"
