+++ typesense/src/document.rs
@@ -3,7 +3,7 @@
 //! In Typesense, documents are each one of the JSON elements that are stored in the collections.
 //! A document to be indexed in a given collection must conform to the schema of the collection.
 //!
-use crate::collection::CollectionSchema;
+use crate::collection_schema::CollectionSchema;
 use serde::{de::DeserializeOwned, Serialize};
 
 /// Trait that should implement every struct that wants to be represented as a Typesense
